% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley}
\alias{barley}
\title{Barley grain yield}
\format{
A data frame with 7395 rows and 3 variables:
\describe{
  \item{X}{X coordinate, in meters}
  \item{Y}{Y coordinate, in meters}
  \item{Yield}{grain yield, in ton per hectare}
}
}
\usage{
barley
}
\description{
A dataset containing Barley grain yield using calibrated commercial
yield monitors mounted on combines equipped with DGPS.
}
\details{
Coordinate reference system is "WGS 84 / UTM zone 20S", epsg:32720
}
\keyword{datasets}
