% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind}
\alias{bind}
\alias{cbind.paar}
\title{Bind outlier condition to an object.}
\usage{
\method{cbind}{paar}(..., deparse.level = 1)
}
\arguments{
\item{...}{objects to bind.}

\item{deparse.level}{integer controlling the construction of labels in
    the case of non-matrix-like arguments (for the default method):\cr
    \code{deparse.level = 0} constructs no labels;\cr
    the default \code{deparse.level = 1} typically and
    \code{deparse.level = 2} always construct labels from the argument
    names, see the \sQuote{Value} section below.}
}
\value{
\code{cbind} called with m.
}
\description{
Bind outlier condition to an object.
}
