% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depurate.R
\name{is_error_update}
\alias{is_error_update}
\title{Auxiliary function for update is_error object in depurate function}
\usage{
is_error_update(is_error, remove_result)
}
\arguments{
\item{is_error}{internal obsect}

\item{remove_result}{internal results from remove_*}
}
\description{
Auxiliary function for update is_error object in depurate function
}
