% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.paarCondition}
\alias{plot.paarCondition}
\title{Plot paar data with condition}
\usage{
\method{plot}{paarCondition}(data, use_ggplot = TRUE)
}
\arguments{
\item{data}{An `sf` object containing a column named `condition`.}

\item{use_ggplot}{Logical. If `TRUE` (default), uses `ggplot2` for plotting 
if the package is installed. If `FALSE`, or if `ggplot2` is not available, 
uses base R plotting.}
}
\value{
If `ggplot2` is used, returns a `ggplot` object invisibly. Otherwise returns `NULL`.
}
\description{
Plots an `paar` data, resulting from `cbind(paar, data)` object colored by a `condition` column. 
The function supports both `ggplot2` and base R graphics.
}
\details{
This function is an S3 method for objects of class `paarCondition`.
The input `data` should be an `sf` object with a `condition` column.
Typically, this object results from combining filtered result data 
from the function \link[paar]{depurate},
using `cbind(paar_data, data)`, where `paar_data` contains the results 
returned by `depurate`, and `data` is the original dataset used in the procedure.
}
