% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.paar}
\alias{print.paar}
\title{Print paar objects}
\usage{
\method{print}{paar}(x, n = 3, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{n}{an integer vector specifying maximum number of rows or
elements to print.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
invisivle object x
}
\description{
Print paar objects
}
