% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.paar}
\alias{print.summary.paar}
\title{Print summarized paar object}
\usage{
\method{print}{summary.paar}(x, digits, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link[base]{print.default}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A data.frame with the summarized condition of the object.
}
\description{
Print summarized paar object
}
