% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depurate.R
\name{remove_border}
\alias{remove_border}
\title{Remove borders}
\usage{
remove_border(x, crs = NULL, buffer, poly_border = NULL)
}
\arguments{
\item{x}{an \code{sf} points object}

\item{crs}{coordinate reference system: integer with the EPSG code,
or character with proj4string to convert coordinates if \code{x} has
longitude/latitude data}

\item{buffer}{\code{numeric} distance in meters to be removed. Negative
values are recommended}

\item{poly_border}{\code{sf} object with one polygon or NULL. Can be
the result of \code{concaveman::concaveman}}
}
\description{
Remove borders
}
\details{
Removes all points from \code{x} that are \code{buffer} meters from
  boundary.
}
