% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depurate.R
\name{remove_outlier}
\alias{remove_outlier}
\title{Removes outliers}
\usage{
remove_outlier(x, y, ylimitmax = NA, ylimitmin = 0, sdout = 3)
}
\arguments{
\item{x}{an \code{sf} points object}

\item{y}{\code{character} with the name of the variable to use for
depuration process}

\item{ylimitmax}{\code{numeric} of length 1 indicating the maximum limit
for the \code{y} variable. If \code{NA} \code{Inf} is assumed}

\item{ylimitmin}{\code{numeric} of length 1 indicating the minimum limit
for the \code{y} variable. If \code{NA} \code{-Inf} is assumed}

\item{sdout}{\code{numeric} values outside the interval
\eqn{mean ± sdout × sdout} values will be removed}
}
\description{
Removes outliers
}
