% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.paar}
\alias{summary.paar}
\title{Summarizing paar objects}
\usage{
\method{summary}{paar}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
An object of class summary.paar (data.frame) with the following columns:
\itemize{
 \item \code{condition} a character vector with the final condition.
 \item \code{n} a numeric vector with the number of rows for each condition.
 \item \code{percentage} a numeric vector with the percentage of rows for each condition.
 }
}
\description{
Summarizing paar objects
}
