% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_k_means.R
\name{fuzzy_k_means}
\alias{fuzzy_k_means}
\title{Fuzzy k-means clustering}
\usage{
fuzzy_k_means(
  data,
  variables,
  number_cluster = 3:5,
  fuzzyness = 1.2,
  distance = "euclidean"
)
}
\arguments{
\item{data}{sf object}

\item{variables}{variables to use for clustering, if missing, all numeric
variables will be used}

\item{number_cluster}{\code{numeric} vector with number of final clusters}

\item{fuzzyness}{A number greater than 1 giving the degree of fuzzification.}

\item{distance}{\code{character} Must be one of the following:
If "euclidean", the mean square error, if "manhattan", the mean
absolute error is computed. Abbreviations are also accepted.}
}
\value{
a list with classification results and indices to select best number of
clusters.
}
\description{
Performs a vectorized fuzzy k-means clustering, this procedure
it is not spatial. The function is almost a wrapper of the function cmeans
from the package e1071. Is intended to be used when `KM-sPC` procedure is not
possible because data set has only 1 variable.
}
