% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_t_test.R
\name{spatial_t_test}
\alias{spatial_t_test}
\title{Modified t test}
\usage{
spatial_t_test(data, variables)
}
\arguments{
\item{data}{\code{sf} data to extract coordinates or two
columns \code{matrix} or \code{data.frame} specifying coordinates.}

\item{variables}{\code{character} vector with column names to perform ttest}
}
\value{
a data.frame with the correlation and p-value for each pair
 of variables
}
\description{
Performs a modified version of the t test to assess the
correlation between spatial processes. See SpatialPack::modified.ttest for details.
}
