% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.paar}
\alias{summary.paar}
\title{Summarizing paar objects}
\usage{
\method{summary}{paar}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summarizing paar objects
}
